requirements.txt
# aws-cdk-lib==2.87.0
aws-cdk-lib==2.92.0
constructs>=10.0.0,<11.0.0
pyyaml
cdk-nag==2.27.202